+++ src/App.js
@@ -5,9 +5,9 @@ import {
   Route,
 } from 'react-router-dom';
 import './App.css';
-import Books from './redux/books/Books';
+import Books from './pages/Books';
 import Navbar from './components/Navbar';
-import Categories from './redux/categories/Categories';
+import Categories from './pages/Categories';
 
 // const store = configureStore();
 function App() {
