+++ src/pages/Categories.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import '../styles/Categories.css';
+
+const Categories = () => (
+  <div className="big-container">
+    <button className="categories-btn" type="button">Check Status</button>
+  </div>
+);
+
+export default Categories;
